fix string to epicsUInt32 conversion via double

+++ modules/database/src/ioc/db/dbConstLink.c
@@ -94,7 +94,7 @@ static long cvt_st_UInt32(const char *from, void *pfield, const dbAddr *paddr)
         status = epicsParseFloat64(from, &dval, &end);
         if (!status &&
             dval >=0 &&
-            dval <= ULONG_MAX)
+            dval <= UINT_MAX)
             *to = dval;
     }
     return status;
